refactor: Extract `WorkflowInstance.ID` into a standalone `WorkflowIdentifier` struct
I've gone back and forth on this one. Back to `WorkflowIdentifier`.
